teensync-backend/requirements.txt
# ── Core Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Database ────────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0

# ── Auth & Security ─────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ── Configuration ───────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ── NLP & Sentiment Analysis ────────────────────────────────────────────────────
nltk==3.9.1
textblob==0.18.0
vaderSentiment==3.3.2
spacy==3.8.3

# ── Machine Learning (Burnout Detection) ────────────────────────────────────────
scikit-learn==1.6.0
numpy==2.2.0
pandas==2.2.3

# ── RAG Pipeline ─────────────────────────────────────────────────────────────────
# sentence-transformers: encodes text into 384-dim vectors (all-MiniLM-L6-v2)
sentence-transformers==3.3.1
# faiss-cpu: fast approximate nearest-neighbor search for retrieval
faiss-cpu==1.9.0

# ── HTTP Client (for optional LLM integration) ──────────────────────────────────
httpx==0.28.1

# ── WebSockets (Peer Support) ───────────────────────────────────────────────────
websockets==14.1

# ── CORS & Middleware ───────────────────────────────────────────────────────────
starlette==0.41.3

# ── Utilities ───────────────────────────────────────────────────────────────────
python-dateutil==2.9.0
email-validator==2.2.0

# ── Deployment ──────────────────────────────────────────────────────────────────
gunicorn==23.0.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
